// src/interfaces/NewVehicle.tsx
// Generated by https://quicktype.io

export interface NewVehicle {
    status: string;
    result: Result;
}

export interface Result {
    code: string;
    msg:  Msg;
}

export interface Msg {
    brand:         string;
    model:         number;
    plates:        string;
    engine_number: string;
    serie:         string;
    id:            string;
}